requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
PyJWT
python-dotenv
pypdf
openai
